Cycles: Automatically create adaptive sampling and viewport passes
Fixes non-working adaptive sampling in the viewport: the missing passes
was the only reason for that.

Now the synchronization code is supposed to store passes in Scene and
never provide them via BufferParams.

The passes are assigned to the BufferParams when dopin session reset.
This way the rendering code has access to proper stride information.
In the future we can remove passes from BufferParams and pre-compute
strides and offsets before passing the BufferParams down the road.

Had to explicitly remove the passes: can not rely on the fact that the
synchronization code overrides passes in scene as this is not the case
for viewport rendering.