Fix fireflies in barbershop due to self intersection
With the new handling of lights as regular intersections, float precision
could lead to self-intersection, causing the light to be counted many times
over and leading to fireflies.

This could be generalized to work for all primitives instead of just lights,
but that's more involved and could be done later. For now this is a simple
isolated change.