Cycles: Move denoising passes to generic passes routines
The change makes it so denoising passes are handled in the same exact
way as "regular" passes are, allowing to get rid of some exceptions in
the code.

The current downside is that if OptiX denoiser is requested and is not
supported by the platform the passes will still be allocated. I'd say
that behavior in this case needs to be revisited: either fallback to
the OpenImageDenoiser (similar to how unsupported GPU Compute falls
back to CPU) or to move support checks in a centralized place. The one
thing which could help here is moving denoiser parameters from Session
to Scene.

The immediate benefit is that the special access to the denoiser passes
in the Buffers is no longer needed.

After this change it will also be possible to save up memory by not
allocating passes which are not needed for the denoiser: for example if
denoiser does not use normals there is no need to allocate pass for
them.