Cycles: Fix and improve behavior of SessionParams::modified
The check around denoising.use seemed to be wrong: seems to be a
mistake of trying to inline check for an effective value of the
new parameters denoiser usage (with device support taken into
account).

Additionally, don't compare denoising parameters if the denoising
is not used. This way it is possible to modify denoiser settings
before enabling it, which is handy in the complex files.