Cycles: Rework passess data access
The idea is to move pass access away from RenderBuffers to a place
where list of passes is known. Most obvious choice is the Session.
A bit of trickery comes from the fact that in theory PathTrace might
have multiple RenderBuffers (in the case of multi-device render).
So the pass access happens via session into the PathTrace.

There is a helper class which pre-calculates various offsets which
can be re-used by updates form multiple devices.