Cycles: Remove background info from Device
There shouldn't be a logic on the device level which will make it
to change behavior based on whether it is an interactive or offline
rendering.

The only current use is in the OpenCL device which initialized
requested features differently for viewport. This code will either
fully fade away or will be redone.