Cycles: Rework denoiser integration
The goal is to allow OptiX denoiser when rendering with CPU or CUDA.

The Denoiser now can request device to be created if none of the
devices configured for rendering is suitable for denoising. This
allowed to remove some tricky logic from BlenderSync.

Another related change is the move of denoiser parameters from Session
to Integrator, allowing to change denoiser settings without doing full
session re-creation.

Known remaining aspects:

- If CUDA is used for rendering try to use same physical device for
  OptiX denoising.

- Optimize data transfer when different Device is used for denoising.

- Would also be nice to not restart rendering when it is only start
  denoising sample is changed.