Cycles: Implement early stopping for adaptive sampling
Stop scheduling path tracing jobs one all pixels did converge.
Since the convergence check is done for the entire big tile it
is not that visible with regular noise level settings and number
of samples. However, still feels like this is proper thing to do,
especially for the idea of gradually decreasing error threshold
for the viewport rendering.

For the GPU there is a single int memory allocated for every
invocation of convergence check kernel. Not sure if it gives
any measurable difference and whether we'd better avoid such
tiny allocations.