Cycles: Fix denoising artifacts with adaptive sampling
Need to make sure denoiser input is has consistent number of samples
for all pixels, otherwise denoiser might see some areas too dark when
they converged.

The output of denoiser needs to be scaled back, so that when combined
pass is scaled with number of samples n display on or pixel buffer
request it gives proper result.

While this is technically slower than the previous state of the branch,
the number of computations is similar to the master branch.

We can optimize some areas: for example by multi-threading OIDN passes
creation.

Additionally, enable handling of all passes for OptiX. Some code was
disabled during development and was never enabled back :(