Cycles: Add active pixels "overlay" option
The purpose of the change is to make it visible which pixels are being
sampled when adaptive sampling is enabled. This is super useful for
development in the area to see which pixels are considered noisy. It
also will be much useful for the upcoming development of the idea of
progressively decreasing noise floor in the viewport.

This is also something what could be useful for lighting artists, TDs,
and other production folks to investigate shots. So in the long run it
will be very nice to make this is an official option.

There are two main stoppers from making the option available:

- The color of "overlay" is hardcoded.
- When navigating in viewport all the pixels are changed, so users are
  "attacked" by this bright fullscreen overlay.

Those are solvable issues, but is not something what is the highest
priority at this time. So, keeping all this in mind the decision on
the interface goes as following:

- This is an option which is only enabled when Developer UI and Cycles
  Debug options are enabled in the User Preferences.

- This is an option in the viewport shading, as it is only affecting
  the viewport.

In the future we will always display this option, regardless of the
user preferences.

The BlenderSession and BlenderSync now are aware of the Developer UI,
which is another good thing I'd say because then the debug panel
handling can happen via the BlenderSession.