Cycles: Pass adaptive sampling threshold to filter kernel
Before the threshold was used from the KernelIntegrator, which makes it
very tricky to modify on runtime. The only other place in the kernel
which was checking for the threshold is passes write code, which should
be possible to only check for existence of passes.

The change moves forward with the allowance of progressively lowering
the noise floor in the viewport render.

So far should be no functional changes.