Cycles: Allow resetting convergence flag of pixel
This way it is possible to resume rendering with lower adaptive
sampling threshold.

Done as an extra argument to the convergence check function instead
of always re-calculating the flag based on the following ideas:

- The reset flag is the same for all threads, so it should not harm
  coherency of GPU threads.

- Reset is not always used, and in cases when it is not used it is
  cheaper to check boolean flag than to calculate the error norm.

So far again should be no functional changes, preparing for the new
development in the viewport.