Cycles: Refactor adaptive sampling filtering work
Move the code to own function and own section of RenderWork.
Preparing for a more comprehensive logic which will progressively
be bringing down noise floor in viewport.

There could be some performance impact due to an extra parallel
for call over the render works, but in practice it should not be
measurable.