Fix compilation error with strict flags
It could be seen as too much of a boiler plate code to explicitly
mention all kernels, but when one does not constantly work with the
related code it is a bit hidden to see where modifications are to
be made when new kernel is added. Having compiler to nag about places
where modification might be needed is personally found to be very
helpful.