Cycles X: Allow adding samples to pixel
The idea is to be able to resume sampling pixel after it "skipped"
some of the scheduled tiles due to convergence.

This is a bit weak concept, but is the most straight forward to
implement. This will be needed for the progressive noise floor
feature.

The current implementation is relying on the fact that the maximum
number of samples will not exceed floating point precision: for
the range of 0 to 2^24 the worst prevision of the single floating
precision is 1. To be more robust for all possible GPUs which might
have precision issues this will be changed in the followup commit.