Cycles X: Experiment with progressively lowering noise in viewport
The idea is make it so in viewport all pixels are uniformly noisy,
and the noise gradually becomes lower and lower. This is like replacing
viewport samples with smarter adaptive sampling.

Currently, rendering starts with noise threshold of 0.4, and once
majority of pixels did converge the noise threshold gets halved, until
the configured threshold is reached.

The final result of viewport rendering should be almost the same as
prior to the patch, but because of extra box filtering on a more sparse
pixel sets there could be some differences. Especially visible with the
active pixels overlay in the case when pixels did not converge to the
final noise floor.

From the benchmarks there seems to be no performance loss.

Differential Revision: https://developer.blender.org/D11088