Fix wrong render with Cycles persistent data
Need to always zero render buffers, otherwise it will have previous
result stored in it, making integration wrong (image will be over
exposed) and will make it so that adaptive sampling will pretend that
it adds samples, causing an early convergence.

Probably the render buffer is to be freed at the end of the render
anyway, but this would need a deeper look. For now just mimicking
what happens in master.