Fix Cycles X crash using a BSDF in a light shader
* Move BSDF feature test so the right offset is applied
* Exclude BSDFs from volume shaders, and volumes from surface shaders,
  to prevent a similar issues.

It would be most efficient to exclude BSDFs entirely from light shaders, but
this is difficult since Cycles does not make a distinction. A shader can be used
by both a Mesh and Light and is compiled only once. We do make a distinction
between surface/volume/displacement shaders, so for that case we can ignore
the node during shader compilation.

Differential Revision: https://developer.blender.org/D11164