Cycles X: add path state compaction for GPU rendering
Before enqueing more work tiles, compact the path states so that all active
states are at the start. This improve coherence by avoiding paths from
different tiles to be interleaved.

It only provides minor performance benefits at this point, and only on
some GPUs. But it can be a useful building block for further optimizations.

Differential Revision: https://developer.blender.org/D11346