Cycles X: Optionally approximate shadow in matte pass
This change makes it so that combined pass (from the render result
point of view) will contain both synthetic objects and shadows when
the Shadow Catcher pass is not enabled.

This brings back behavior of the shadow catcher, possibly bringing
back some of workflows when synthetic object have to have their
shadows in the same pass. It also makes it so perceptually the
shadow catcher always work.

The approximation is a black transparency (without color), although
is more accurate than in previous implementation of the shadow
catcher.

Currently is only implemented for the final render. The viewport
needs more investigation about whether its possible to have preview
of the final accurate shadow catcher.

Differential Revision: https://developer.blender.org/D11232