Cycles X: Generalize storage of offsets in BufferParams
This change replaces storage of a hand-picked passes with a more
generic storage. Simplifies code in the BufferParams and makes
data available in all areas which need to access those offsets.

A bit of a duplication with KernelFilm, but not all areas have
access to the KernelFilm. Film can not be used either because
it is not guaranteed to be in sync with the BufferParams. So in
a way now all the knowledge about what is stored and how to be
accessed in the Buffer is available via BufferParams.

Simplifies arguments passed to the PassAccessor.
The Film is still needed to be passed there to know whether
shadow from the shadow catcher pass is to be averaged and added
to the combined pass or not.

Currently no functional changes, just preparation for an upcoming
deeper refactor of the PassAccessor.

Differential Revision: https://developer.blender.org/D11437