Cycles X: Move pass accessor to integrator folder
This is where it is used the most and will be used even more in the
future.

No functional changes, preparing for an upcoming generalization of
the class. Eventually this class will take over the kernel film
convert kernel, for which DeviceQueue will be needed, so it makes
sense to move the pass accessor to the integrator folder

Differential Revision: https://developer.blender.org/D11476