Cycles X: Multi-thread CPU pass accessor
Is not very measurable for the combined pass, but for shadow catcher
matte with approximated shadow gives about 30% speedup of the display
update.

Surely, this is only fraction of the overall render time, but is
still nice to have it. Makes it easier to re-use for the viewport
update as well.

Differential Revision: https://developer.blender.org/D11479