Fix Cycles-X GPU rendering and OIDN denoiser
The issue was introduced in the initial baking support commit:
the `RenderBuffers::zero()` does not modify the CPU-size data,
so sequential call to `RenderBuffers::copy_to_device()` will
copy old data to the device.

Solved by not copying data to the device if there is no data
read, which is good thing to do anyway to avoid double data
copy to the device when no new data was provided.