Reduce kernel type boiler plate code in Cycles X
This is kind of a step back from an initial idea of making it easier
to find where kernel type is handled. It worked fine when number of
kernels was low. Now there are so many kernels and only so few of
them require special case that being explicit is actually in a way.

Now there corner cases are really handled as a corner cases, and there
is a note about where to look for those written next to the kernel
enum definition.

Differential Revision: https://developer.blender.org/D11571