Cycles X: De-duplicate shadow pass accessor kernel
The downside is that there is one extra if statement per pixel, but it
is probably unmeasurable because of other calculations involved. Could
not run actual numbers because it is not a function which is run often.

The upside is that it opens possibilities for ease support of passes
with optional alpha channel. For example, it will become useful for
shadow catcher pass denoising.

We can also now implement passes with divider for both RGB and RGBA
destinations and get rid of an exception in the pass Destination
constructor.

Differential Revision: https://developer.blender.org/D11573