Cycles X: Initial implementation of shadow catcher denoising
Currently only implemented for OIDN denoiser.

Almost straightforward. The tricky part is the albedo pass: using the
real albedo causes unexpected discontinuity of the shadow catcher based
on the albedo of the surface the shadow is cast on. Worked around by
passing a fake albedo values. The downside is that it requires a full
frame of constant values. Not sure if there is a way around it.

Differential Revision: https://developer.blender.org/D11585