Cycles X: Implement float support in PassAccessorGPU
Rather straightforward implementation, with some code-generation
macro which takes care of some boiler plate code.

Unfortunately, clang-format does some weird decision in the
`kernel_as_string ` function. Attempt to disable clang-format via
comment makes the comment to be weirdly indented. So this part
is left-as is.

Similar to the PassAccessorCPU there is no overlays applied to
the float result.

Tested with an upcoming refactor of the OptiX denoiser which
uses pass accessor.

No functional changes so far.