Cycles X: Initial support for the constant time rendering
Allows to limit render time, so that rendering will either happen
until all samples are finished, or the given time limit is reached.

Currently is only implemented for background and headless renders,
as those seems to benefit the most. In the viewport would still
like to investigate idea of keeping rendering until desired noise
floor is reached before adding extra settings there.

The usecases for the non-interactive (background, headless) renders
includes:

- Rendering a shot previews for movies.
- Ensuring some "bogus" frames do not stall the render farm.
- Allows to implement benchmark which will work reliable for both
  very powerful and weak hardware configurations (as an example,
  top-end GPUs and mobile GPUs) without requiring to wait forever
  on a non-powerful hardware.

One of the tricky parts which is not solved in this version yet is
support of constant time rendering with multiple big tiles. However,
support of big tiles will be an addition on top of the existing code
as some sort of time limit would need to be enforced for every tile
(in other words, support of constant time render for multiple tiles
will be implemented on a higher level).

Differential Revision: https://developer.blender.org/D11526