Cycles X: Reduce memory usage of OptiX denoiser
The idea is to perform denoising in the render buffer's denoised pass
without allocating temporary color.

This optimization covers all scenarios of when OptiX denoiser is used:
single- and multi- device rendering.

The change is a bit bigger than it could be from the minimal functional
point of view, but kept running into some issues and seemed helpful to
make certain parts more clear.

Differential Revision: https://developer.blender.org/D11847