Cycles X: Implement path compaction for shadow catcher
The demo file is BMW27 with the ground set as a shadow catcher.
The observed performance improvement is about 5% on RTX5000.

The general idea is to schedule new tiles in a way that we always
leave space for the shadow catcher. Roughly, we first schedule 50%
of path states from the maximum number of paths, then 25% and so on.

Summary of changes:

- Replace constant offset of shadow catcher state with an atomically
  incrementing index.

- Add new kernel to count number of states which can still spit.

  Could experiment with some atomics so that path split decreases a
  value, so does path termination, and increase it when new paths
  are added. Not sure this will give better performance.

- Remove terminated paths kernel from scheduling.
  The paths are compacted, so we know they are in the beginning of
  the array.

Differential Revision: https://developer.blender.org/D11932