Cycles X: Tweak max number of states seen by tile scheduler
This is required for shadow catchers to make it so the tile scheduler
gives work which can fir into the number of allowed camera rays. Use
a smaller value from the maximum number of states to prepare code for
state compaction of re-scheduling for the shadow catcher.

Interestingly, this has positive effect on regular rendering here with
RTX 5000:
```
                              new                           cycles-x
bmw27.blend                   12.445                        12.2104
classroom.blend               24.4949                       24.4508
pabellon.blend                11.3019                       11.4407
monster.blend                 13.409                        13.4491
barbershop_interior.blend     18.6601                       18.8364
junkshop.blend                26.3212                       27.051
pvt_flat.blend                22.7389                       22.9345
```

For the future development we might try to make it so tile scheduler
gives smaller tiles with smaller number of samples, rely on the path
work GPU to request as many tiles as fit into the path states. Need
to be careful though, because there are downsides in terms of memory
bandwidth to pass works tiles to the init_from kernels.