Remove accurate cryptomatte implementation
Affects both Cycles and EEVEE.

The motivation:

- For EEVEE it is a performance penalty to do GPU transfers for every
  sample rendered.

- With big tile rendering in Cycles the accurate mode can have a very
  high memory footprint.

- Feature parity between CPU and GPU: the GPU code paths should provide
  results which are usable by production.

If there are cases when the accurate is really a must-have we'd need to
have a demo file and investigate how it can be supported on both CPU
and GPU.

Note that for comparison of accurate vs. non-accurate in the regular
Cycles should be done very carefully: CPU does not implement sorting
for non-accurate mode, so doesn't OptiX; and GPU does not support
accurate modes; CUDA doesn't implement sorting properly when adaptive
sampling is involved (what a mess!).

Differential Revision: https://developer.blender.org/D11975