Fix OpenGL context usage in Cycles X
Activating nullptr context is not a valid way to release the active
context, there is a dedicated function for this.

This also showed a logical error in the buffer mapping: the buffer
unmap is not supposed to release current OpenGL context but should
instead leave it active for a possible subsequent calls to the GPU
update. It worked before because the context as never released due
to wrong API usage.

Would say that the WM OpenGL API should have an assert to catch
such case, but this better to happen in the master branch.

Noticed while doing a further development related to GPUDisplay and
watching return value of GHOST_ActivateOpenGLContext.

Differential Revision: https://developer.blender.org/D11989