Fix shadow catcher casting unnecessary shadows in some cases
When both the render with and without synthetic objects results in a black
pixel and we get 0/0, we previously would assume that area is fully shadowed.
Now assume there is no shadow.

The correct result is unknown in this case. But for the case where an adjacent
pixel has 0.001/0.001, this avoids a discontinuity.