Fix missing pre-filtering execute in the OIDN Cycles X
The `execute()` call got missed in some cleanup of the initial hacky
implementation and indicatives of it did not get proper attention.

The pre-filter takes about 0.5sec per pass on i9-11900k Full-HD image
which indeed makes it not suitable for the preview render.

The file used to verify this change: F10247300

Result:
|No filter|With filter|
|{F10247303}|{F10247304}|