Fix dead-lock when cancelling viewport render in Cycles X
This change is supposed to fix a hard-to-reproduce bug when Blender
gets stuck when quitting while viewport render ins going.

Ensure consistency in the check for the pause, so that it is checked
before asking to wait the condition variable. Otherwise it was possible
to miss the notification and not react to the pause state properly.