Fix 100% core usage in viewport in Cycles X
Is a regression since the previous fix for the dead-lock on exit
(was done in the revision 24458a1e7a0).

We need to do `pause_cond_.wait()` when we've run out work to be done.

The new code does check for both states: pause and new work added.

Annoyingly, had to introduce an extra flag which we can check prior
to `wait()`. Without this we have nothing to check whether wait is
needed or not.

The change is against Cycles X, but would also need to be ported
to master.

Differential Revision: https://developer.blender.org/D12148